refactor: remove @nuxt/common
