--- packages/common/data/modern-browsers.json
@@ -1,12 +1,0 @@
-{
-  "Edge": "16",
-  "Firefox": "60",
-  "Chrome": "61",
-  "Chrome Headless": "61",
-  "Chromium": "61",
-  "Safari": "10.1",
-  "Opera": "48",
-  "Yandex": "18",
-  "Vivaldi": "1.14",
-  "Mobile Safari": "10.3"
-}
