--- packages/common/package.js
@@ -1,3 +1,0 @@
-export default {
-  build: true
-}
