--- packages/common/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@nuxt/common",
-  "version": "2.3.4",
-  "repository": "nuxt/nuxt.js",
-  "license": "MIT",
-  "files": [
-    "dist"
-  ],
-  "main": "dist/common.js",
-  "dependencies": {
-    "@nuxt/utils": "2.3.4",
-    "consola": "^2.3.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
