--- packages/common/src/index.js
@@ -1,2 +1,0 @@
-export { default as Hookable } from './hookable'
-export { default as ModernBrowsers } from '../data/modern-browsers.json'
